Refactored imports and added poetry version bump
