+++ .pre-commit-config.yaml
@@ -1,17 +1,44 @@
 repos:
+
+  - repo: https://github.com/Tesla2000/TupleNamer
+    rev: v0.0.3-alpha
+    hooks:
+     - id: tuple_namer
+       stages: [pre-push]
+  - repo: https://github.com/Tesla2000/EmptyTypeCheckingRemover
+    rev: v0.0.1
+    hooks:
+     - id: empty_type_checking_remover
+       stages: [pre-commit]
   - repo: https://github.com/psf/black
     rev: 25.1.0
     hooks:
       - id: black
         args: ["--preview", "--line-length", "79"]
+  - repo: https://github.com/Tesla2000/import-rules-enforcer
+    rev: 'v0.0.12'
+    hooks:
+      - id: import_rules_enforcer
+  - repo: https://github.com/Tesla2000/temp-reorder-python-imports
+    rev: v0.0.1
+    hooks:
+      - id: reorder-python-imports
+        args: [
+          --py37-plus,
+          --py39-plus,
+          --retain-pre-import,
+          'True',
+          --application-directories,
+          "cm_client",
+        ]
+  - repo: https://github.com/PyCQA/autoflake
+    rev: v2.3.1
+    hooks:
+      - id: autoflake
+        args: [ --remove-all-unused-imports, --in-place ]
   - repo: https://github.com/pycqa/flake8
     rev: '7.3.0'
     hooks:
       - id: flake8
         args: [ "--ignore=E203,W503,E501,E704" ]
-  - repo: https://github.com/pre-commit/mirrors-isort
-    rev: v5.10.1
-    hooks:
-      - id: isort
-        args: [ "--profile", "black", "--line-length", "79" ]
 default_stages: [commit]
