+++ dependify/decorators/_injected/_get_annotations.py
@@ -1,13 +1,22 @@
 import sys
-from typing import Callable, Dict, Type, TypeVar, Union
+from typing import Callable
+from typing import Dict
+from typing import Type
+from typing import TypeVar
+from typing import Union
 
 class_type = TypeVar("class_type", bound=type)
 
-def get_annotations(class_or_constructor: Union[Type, Callable]) -> Dict[str, Type]:
+
+def get_annotations(
+    class_or_constructor: Union[Type, Callable],
+) -> Dict[str, Type]:
     if not isinstance(class_or_constructor, type):
         return class_or_constructor.__annotations__
     class_annotations = {}
-    for ancestor in reversed(class_or_constructor.__mro__[:-1]):  # remove object
+    for ancestor in reversed(
+        class_or_constructor.__mro__[:-1]
+    ):  # remove object
         if sys.version_info >= (3, 10, 0):
             annotations = ancestor.__annotations__
         else:
